Refactor: create separate folder for modules
