+++ main.html
@@ -87,11 +87,11 @@
                 <input type="button" id="redraw" value="Redraw"/>
             </div>
         </div>
-        <script type="module" src="./js/utils.js"></script>
-        <script type="module" src="./js/complex.js"></script>
-        <script type="module" src="./js/fractal.js"></script>
-        <script type="module" src="./js/frame.js"></script>
-        <script type="module" src="./js/image.js"></script>
+        <script type="module" src="./js/modules/utils.js"></script>
+        <script type="module" src="./js/modules/complex.js"></script>
+        <script type="module" src="./js/modules/fractal.js"></script>
+        <script type="module" src="./js/modules/frame.js"></script>
+        <script type="module" src="./js/modules/image.js"></script>
         <script type="module" src="./js/main.js"></script>
     </body>
 </html>
